+++ NEWS.md
@@ -1,3 +1,4 @@
 # lfl 2.0.1
 * added support for percentual quantifiers in compose()
 * documentation spell-check
+* the result of algebra() is now an S3 class, implemented print.algebra()